Revert "refactor: move test_util.ts"
This reverts commit cdaa729a08b404e57faa8bb55bba3eb715be1913.
