+++ tests/unit/chmod_test.ts
@@ -4,7 +4,7 @@ import {
   assertEquals,
   assertRejects,
   assertThrows,
-} from "../util/test_util.ts";
+} from "./test_util.ts";
 
 Deno.test(
   {
